% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_rv_self.R
\name{discrete_rv_self}
\alias{discrete_rv_self}
\title{function to generate discrete uniform random variables for initial random design points in ForLion}
\usage{
discrete_rv_self(n, xlist)
}
\arguments{
\item{n}{number of discrete random variables}

\item{xlist}{list of levels for variables to be generated}
}
\value{
list of discrete uniform random variables
}
\description{
function to generate discrete uniform random variables for initial random design points in ForLion
}
\examples{
n=3 #three discrete random variables
xlist=list(c(-1,1),c(-1,1),c(-1,0,1)) #two binary and one three-levels
discrete_rv_self(n, xlist)


}
