% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dprime_func_self.R
\name{dprime_func_self}
\alias{dprime_func_self}
\title{function to calculate du/dx in the gradient of d(x, Xi), will be used in ForLion_MLM_func() function, details see Appendix C in Huang, Li, Mandal, Yang (2024)}
\usage{
dprime_func_self(
  xi,
  bvec,
  h.func,
  h.prime,
  inv.F.mat,
  Ux,
  link = "continuation",
  k.continuous
)
}
\arguments{
\item{xi}{a vector of design point}

\item{bvec}{parameter of the multinomial logistic regression model}

\item{h.func}{function, is used to transfer xi to model matrix (e.g. add interaction term, add intercept)}

\item{h.prime}{function, is used to find dX/dx}

\item{inv.F.mat}{inverse of F_Xi matrix, inverse of fisher information of current design w/o new point}

\item{Ux}{U_x matrix in the algorithm, get from Fi_MLM_func() function}

\item{link}{multinomial link function, default is"continuation", other choices "baseline", "cumulative", and "adjacent"}

\item{k.continuous}{number of continuous factors}
}
\value{
dU/dx in the gradient of sensitivity function d(x, Xi)
}
\description{
function to calculate du/dx in the gradient of d(x, Xi), will be used in ForLion_MLM_func() function, details see Appendix C in Huang, Li, Mandal, Yang (2024)
}
