% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EW_dprime_func_self.R
\name{EW_dprime_func_self}
\alias{EW_dprime_func_self}
\title{function to calculate dEu/dx in the gradient of d(x, Xi), will be used in EW_ForLion_MLM_func() function}
\usage{
EW_dprime_func_self(
  xi,
  bvec_matrix,
  h.func,
  h.prime,
  inv.F.mat,
  EUx,
  link = "continuation",
  k.continuous
)
}
\arguments{
\item{xi}{a vector of design point}

\item{bvec_matrix}{the matrix of the bootstrap parameter values of beta}

\item{h.func}{function, is used to transfer xi to model matrix (e.g. add interaction term, add intercept)}

\item{h.prime}{function, is used to find dX/dx}

\item{inv.F.mat}{inverse of F_Xi matrix, inverse of the Expectation of fisher information of current design w/o new point}

\item{EUx}{EU_x matrix in the algorithm, get from EW_Fi_MLM_func() function}

\item{link}{link multinomial link function, default is"continuation", other choices "baseline", "cumulative", and "adjacent"}

\item{k.continuous}{number of continuous factors}
}
\value{
dEU/dx in the gradient of sensitivity function d(x, Xi)
}
\description{
function to calculate dEu/dx in the gradient of d(x, Xi), will be used in EW_ForLion_MLM_func() function
}
