% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu1_cauchit_self.R
\name{nu1_cauchit_self}
\alias{nu1_cauchit_self}
\title{Function to calculate first derivative of nu function given eta for cauchit link}
\usage{
nu1_cauchit_self(x)
}
\arguments{
\item{x}{vector of eta, eta=X*beta}
}
\value{
the first derivative of nu function given eta for cauchit link
}
\description{
Function to calculate first derivative of nu function given eta for cauchit link
}
\examples{
eta = c(1,2,3,4)
nu1_cauchit_self(eta)

}
