% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu1_logit_self.R
\name{nu1_logit_self}
\alias{nu1_logit_self}
\title{function to calculate the first derivative of nu function given eta for logit link}
\usage{
nu1_logit_self(x)
}
\arguments{
\item{x}{vector of eta, eta=X*beta}
}
\value{
the first derivative of nu function given eta for logit link
}
\description{
function to calculate the first derivative of nu function given eta for logit link
}
\examples{
eta = c(1,2,3,4)
nu1_logit_self(eta)


}
