% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu2_loglog_self.R
\name{nu2_loglog_self}
\alias{nu2_loglog_self}
\title{function to calculate the second derivative of nu function given eta for loglog link}
\usage{
nu2_loglog_self(x)
}
\arguments{
\item{x}{vector of eta, eta=X*beta}
}
\value{
the second derivative of nu function for loglog link
}
\description{
function to calculate the second derivative of nu function given eta for loglog link
}
\examples{
eta = c(1,2,3,4)
nu2_loglog_self(eta)

}
