% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu2_probit_self.R
\name{nu2_probit_self}
\alias{nu2_probit_self}
\title{function to calculate the second derivative of nu function given eta for probit link}
\usage{
nu2_probit_self(x)
}
\arguments{
\item{x}{vector of eta, eta=X*beta}
}
\value{
the second derivative of nu function for probit link
}
\description{
function to calculate the second derivative of nu function given eta for probit link
}
\examples{
eta = c(1,2,3,4)
nu2_probit_self(eta)

}
