% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_cauchit_self.R
\name{nu_cauchit_self}
\alias{nu_cauchit_self}
\title{function to calculate w = nu(eta) given eta for cauchit link}
\usage{
nu_cauchit_self(x)
}
\arguments{
\item{x}{a list of eta = X*beta}
}
\value{
diagonal element of W matrix which is nu(eta)
}
\description{
function to calculate w = nu(eta) given eta for cauchit link
}
\examples{
eta = c(1,2,3,4)
nu_cauchit_self(eta)


}
