% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_identity_self.R
\name{nu_identity_self}
\alias{nu_identity_self}
\title{Function to calculate w = nu(eta) given eta for identity link}
\usage{
nu_identity_self(x)
}
\arguments{
\item{x}{Numeric vector of eta, eta = X*beta.}
}
\value{
A numeric vector representing the diagonal elements of the W matrix (nu(eta)).
}
\description{
Function to calculate w = nu(eta) given eta for identity link
}
\examples{
eta = c(1,2,3,4)
nu_identity_self(eta)

}
