% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_logit_self.R
\name{nu_logit_self}
\alias{nu_logit_self}
\title{function to calculate w = nu(eta) given eta for logit link}
\usage{
nu_logit_self(x)
}
\arguments{
\item{x}{vector of eta, eta=X*beta}
}
\value{
diagonal element of W matrix which is nu(eta)
}
\description{
function to calculate w = nu(eta) given eta for logit link
}
\examples{
eta = c(1,2,3,4)
nu_logit_self(eta)


}
