% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_sol_J3.R
\name{polynomial_sol_J3}
\alias{polynomial_sol_J3}
\title{functions to solve 2th order polynomial function given coefficients}
\usage{
polynomial_sol_J3(c0, c1, c2)
}
\arguments{
\item{c0}{constant coefficient of polynomial function}

\item{c1}{coefficient of 1st order term}

\item{c2}{coefficient of 2nd order term}
}
\value{
sol   the 2 solutions of the polynomial function
}
\description{
functions to solve 2th order polynomial function given coefficients
}
\examples{
polynomial_sol_J3(-2,-3, 1)

}
