% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial_sol_J4.R
\name{polynomial_sol_J4}
\alias{polynomial_sol_J4}
\title{functions to solve 3th order polynomial function given coefficients}
\usage{
polynomial_sol_J4(c0, c1, c2, c3)
}
\arguments{
\item{c0}{constant coefficient of polynomial function}

\item{c1}{coefficient of 1st order term}

\item{c2}{coefficient of 2nd order term}

\item{c3}{coefficient of 3rd order term}
}
\value{
sol   the 3 solutions of the polynomial function
}
\description{
functions to solve 3th order polynomial function given coefficients
}
\examples{
polynomial_sol_J4(0,9,6,1)

}
