% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.design_output.R
\name{print.design_output}
\alias{print.design_output}
\title{Print Method for Design Output from ForLion Algorithms}
\usage{
\method{print}{design_output}(x, ...)
}
\arguments{
\item{x}{An object of class `design_output`.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns `x`.
}
\description{
Custom print method for a list containing design information.
}
