% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.list_output.R
\name{print.list_output}
\alias{print.list_output}
\title{Print Method for list_output Objects}
\usage{
\method{print}{list_output}(x, ...)
}
\arguments{
\item{x}{An object of class `list_output`.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns `x` (the input object).
}
\description{
Custom print method for objects of class `list_output`.
}
