% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd_inverse.R
\name{svd_inverse}
\alias{svd_inverse}
\title{SVD Inverse Of A Square Matrix}
\usage{
svd_inverse(x)
}
\arguments{
\item{x}{the matrix for calculation of inverse}
}
\value{
the inverse of the matrix x
}
\description{
This function returns the inverse of a matrix using singular value decomposition.
If the matrix is a square matrix, this should be equivalent to using the solve function.
If the matrix is not a square matrix, then the result is the Moore-Penrose pseudo inverse.
}
\examples{
x = diag(4)
svd_inverse(x)

}
